feat: Refactor OnBoarding Screen and reach on it once
