+++ lib/ui/splash/initial_screen.dart
@@ -1,0 +1,43 @@
+import 'package:flutter/material.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+import 'package:track_fit/ui/auth/view/auth_screen.dart';
+import 'package:track_fit/ui/onboarding/view/onboarding_screen.dart';
+
+class InitialScreen extends StatefulWidget {
+  const InitialScreen({super.key});
+
+  @override
+  State<StatefulWidget> createState() => _InitialScreenState();
+}
+
+class _InitialScreenState extends State<InitialScreen> {
+  Widget? _startScreen;
+
+  @override
+  void initState() {
+    super.initState();
+    _checkFirstLaunch();
+  }
+
+  Future<void> _checkFirstLaunch() async {
+    final prefs = await SharedPreferences.getInstance();
+    final hasSeenOnboarding = prefs.getBool('hasSeenOnboarding') ?? false;
+
+    if (!hasSeenOnboarding) {
+      await prefs.setBool('hasSeenOnboarding', true);
+      _startScreen = const OnboardingScreen();
+    } else {
+      _startScreen = const AuthScreen();
+    }
+
+    setState(() {});
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    if (_startScreen == null) {
+      return const Scaffold(body: Center(child: CircularProgressIndicator()));
+    }
+    return _startScreen!;
+  }
+}
