+++ pubspec.yaml
@@ -41,6 +41,7 @@ dependencies:
   path: ^1.9.1
   path_provider: ^2.1.5
   flutter_riverpod: ^2.6.1
+  shared_preferences: ^2.5.3
 
 dev_dependencies:
   flutter_test:
